+++ packages/url_launcher/url_launcher/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 6.3.3
 
+* Fixes `supportsCloseForLaunchMode` reporting launch support instead of close
+  support.
 * Updates minimum supported SDK version to Flutter 3.41/Dart 3.11.
 * Updates README to reflect currently supported OS versions for the latest
   versions of the endorsed platform implementations.